CWBackPixel tests
The problem is that when there's no background, the handling is not defined
If pixmap is set but there's no pixemap defined (via feh, etc) when you
switch to an empty workspace it shows glitches from prior workspace
If a background_pixel is set, it overrides the pixmap.
If a pixmap is present (pixel not set) evereything works fine.